# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XConnect API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/inventory/transfer": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create a P2P inventory transfer request",
                "operationId": "523da4ddf03b3502e9fc7fd6a0f682fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "receiver_id",
                                    "items"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "description": "Technician ID to receive items",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "notes": {
                                        "description": "Optional notes for transfer",
                                        "type": "string",
                                        "example": "Extra connectors for upcoming installs",
                                        "nullable": true
                                    },
                                    "items": {
                                        "description": "Items to transfer",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "item_id": {
                                                    "description": "Inventory item ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "description": "Quantity to transfer",
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer request created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error (insufficient stock or invalid receiver)"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory/transfer/incoming": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List pending incoming inventory transfers",
                "operationId": "441d3aeaf1d0185a7f15848a0ea6c623",
                "responses": {
                    "200": {
                        "description": "List of pending incoming transfers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "sender_id": {
                                                "type": "integer"
                                            },
                                            "sender_name": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string",
                                                "enum": [
                                                    "pending",
                                                    "accepted",
                                                    "rejected"
                                                ]
                                            },
                                            "notes": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "item_id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "quantity": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory/transfer/{id}/accept": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Accept an incoming inventory transfer",
                "operationId": "27c6658322aaf90196dfabda869fb761",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer accepted successfully"
                    },
                    "403": {
                        "description": "Not the recipient of this transfer"
                    },
                    "404": {
                        "description": "Transfer not found"
                    },
                    "422": {
                        "description": "Transfer not pending or sender insufficient stock"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/inventory/transfer/{id}/reject": {
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Reject an incoming inventory transfer",
                "operationId": "2bcd46bcd6028dc6ac3cf7ac0c77b153",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transfer rejected successfully"
                    },
                    "403": {
                        "description": "Not the recipient of this transfer"
                    },
                    "404": {
                        "description": "Transfer not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Use Sanctum bearer token from /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}